Split integration tests into separate files
