--- test/integration/my_first_test.js
@@ -1,241 +1,0 @@
-'use strict';
-
-var webdriver = require('selenium-webdriver');
-
-var filter = require('./util/filter');
-
-describe('homepage', function() {
-  var driver, selectors;
-
-  beforeEach(function() {
-    driver = this.driver;
-    selectors = this.selectors;
-  });
-
-  it('displays the header', function() {
-    return this.driver.findElement(webdriver.By.css(selectors.homeHeader)).then(function(head) {
-      return head.getText().then(function(headText) {
-        assert.equal(headText, 'Georgia Legislative Navigator');
-      });
-    });
-  });
-
-  describe('members index', function() {
-
-    beforeEach(function() {
-      return driver.findElement(webdriver.By.css(selectors.nav.members))
-        .then(function(membersElement) {
-          return membersElement.click();
-        })
-        .then(function() {
-          return driver.isElementPresent(webdriver.By.css(selectors.layouts.members.region));
-        });
-    });
-
-    it('lists all members from both chambers when the user clicks on the "members" element', function() {
-      return driver.findElements(webdriver.By.css(selectors.layouts.members.thumbnail))
-        .then(function(memberEls) {
-          assert.equal(memberEls.length, 236);
-        });
-    });
-
-    describe('member page', function() {
-      beforeEach(function() {
-        var firstMemberThumbnail;
-
-        // TODO: Conditionally increase test timeout only when web requests
-        // are not being mocked.
-        this.timeout(10000);
-
-        return driver
-          .findElement(webdriver.By.css(selectors.layouts.members.thumbnail))
-          .then(function(el) {
-            firstMemberThumbnail = el;
-            return driver.wait(function() {
-              return firstMemberThumbnail.isDisplayed();
-            });
-          }).then(function() {
-            // TODO: Remove this workaround when application is fixed
-            // to not scroll in response to the "members" route.
-            return driver.executeScript(function() {
-              scrollTo(0, 0);
-            });
-          }).then(function() {
-            return firstMemberThumbnail.click();
-          })
-          .then(function() {
-            return driver.wait(function() {
-              return driver.isElementPresent(
-                webdriver.By.css(selectors.layouts.member.region)
-              );
-            });
-          });
-      });
-
-      it('Bills Sponsored area expands when link is clicked', function() {
-        return driver
-          .findElement(webdriver.By.css(selectors.layouts.member.accordian))
-          .then(function(el) {
-            return el.click();
-          })
-          .then(function() {
-            return driver.wait(function() {
-              return driver.findElement(webdriver.By.css(selectors.layouts.member.billList))
-                .then(function(el) {
-                  return el.isDisplayed();
-                });
-            });
-          });
-      });
-    });
-
-    describe('filtering', function() {
-      this.timeout(8000);
-
-      it('republicans', function() {
-        return filter(driver, 'republicans').then(function(result) {
-          assert.equal(result.republicans, 157);
-          assert.equal(result.democrats, 0);
-          assert.equal(result.independents, 0);
-        });
-      });
-
-      it('democrats', function() {
-        return filter(driver, 'democrats').then(function(result) {
-          assert.equal(result.republicans, 0);
-          assert.equal(result.democrats, 78);
-          assert.equal(result.independents, 0);
-        });
-      });
-
-      it('independents', function() {
-        return filter(driver, 'independents').then(function(result) {
-          assert.equal(result.republicans, 0);
-          assert.equal(result.democrats, 0);
-          assert.equal(result.independents, 1);
-        });
-      });
-
-      it('democrats in the senate', function() {
-        return filter(driver, 'senate').then(function() {
-          return filter(driver, 'democrats');
-        }).then(function(result) {
-          assert.equal(result.republicans, 0);
-          assert.equal(result.democrats, 18);
-          assert.equal(result.independents, 0);
-        });
-      });
-
-      it('republicans in the house', function() {
-        return filter(driver, 'house').then(function() {
-          return filter(driver, 'republicans');
-        }).then(function(result) {
-          assert.equal(result.republicans, 119);
-          assert.equal(result.democrats, 0);
-          assert.equal(result.independents, 0);
-        });
-      });
-    });
-  });
-
-  describe('bills index', function(){
-    beforeEach(function() {
-      return driver.findElement(webdriver.By.css(selectors.nav.bills))
-        .then(function(billsElement) {
-          return billsElement.click();
-        })
-        .then(function() {
-          return driver.isElementPresent(webdriver.By.css(selectors.layouts.bills.region));
-        });
-    });
-
-    it('lists all bill categories when user clickes on bills homepage element', function() {
-      return driver.findElements(webdriver.By.css(selectors.layouts.bills.thumbnail))
-        .then(function(billCatEls) {
-          assert.equal(billCatEls.length, 10);
-        });
-    });
-
-    describe('search bills from bill category page', function(){
-
-      it('searches for HB1', function() {
-        return driver.findElement(webdriver.By.css(selectors.layouts.bills.search.searchform))
-          .then(function(billSearch) {
-            return billSearch.click()
-              .then(function() {
-                return billSearch.sendKeys('1');
-              });
-          })
-          .then(function() {
-            return driver.findElement(webdriver.By.css(selectors.layouts.bills.search.searchbutton))
-              .then(function(button) {
-                return button.click();
-              })
-              .then(function() {
-                return driver.isElementPresent(webdriver.By.css(selectors.layouts.bills.billpage));
-              });
-          });
-      });
-
-      describe('bill search', function() {
-        var billId = '1151';
-
-        it('correctly returns requested house bill', function() {
-          return driver.findElement(
-            webdriver.By.css(selectors.layouts.bills.search.documentType)
-          ).then(function(selectElement) {
-            return selectElement.sendKeys('house bills');
-          }).then(function() {
-            return driver.findElement(
-              webdriver.By.css(selectors.layouts.bills.search.searchform)
-            );
-          }).then(function(numberInput) {
-            return numberInput.sendKeys(billId)
-              .then(function() {
-                return numberInput.submit();
-              });
-          }).then(function() {
-            return driver.wait(function() {
-              return driver.isElementPresent(
-                webdriver.By.css(selectors.layouts.bill.region)
-              );
-            }, 3000);
-          }).then(function() {
-            return driver.findElement(
-              webdriver.By.css(selectors.layouts.bill.title)
-            );
-          }).then(function(element) {
-            return element.getText();
-          }).then(function(text) {
-            assert(
-              text.indexOf(billId) > -1, 'Expected bill ID present on page.'
-            );
-          });
-        });
-
-        it('does not return house bills when searching senate bills');
-      });
-
-      it('every category displays a positive number of watched bills', function() {
-        var countRe = /\b(\d+)\s*bills?\b/i;
-
-        return driver.findElements(
-          webdriver.By.css(selectors.layouts.bills.thumbnail)
-        ).then(function(elements) {
-          assert.equal(elements.length, 10);
-
-          return webdriver.promise.all(elements.map(function(element) {
-            return element.getText();
-          })).then(function(texts) {
-            texts.forEach(function(text) {
-              var match = text.match(countRe);
-
-              assert(match, 'Displays some number of bills');
-              assert(parseInt(match[0], 10) > 0, 'Displays a positive number of bills');
-            });
-          });
-        });
-      });
-    });
-  });
-});
